fix bug in arg handling around 'delete' button in directory view
the code composing the form providing the 'delete' button in a dir
view was broken in that it tried to put some of the arguments into
the url query, rather than as the form's post args.  worse, nevow
was kind enough to escape the querystring making it invalid.

darcs-hash:20070712234654-05425-2f5af18cfb49a1fd2512d7e6d31521cc7ff72931.gz