foolscap: change using-foolscap.xhtml to say "public key" instead of "private key"
In the first instance, I hope that the tubid is generated from the hash of the public key and not the hash of the private key!
In the latter instances, the phrase "public key" as in "public key certificate" is more common.

darcs-hash:20070811164503-92b7f-3743881f9027f15240ce14bce7e44f06282be645.gz