webish: look for when_done= in POST fields as well as queryargs. Closes #101.
We need to look in the fields because that's how we build the mkdir/upload
forms. Without this, uploading or creating directories would leave us on a
page that had just a URI, instead of something actually useful to a human.

darcs-hash:20070814004502-e01fd-2632395c076b686af19f6e1d146f780ebe44f5a2.gz