refactor reading of configuration files in client.py
This makes it so that an optional file which is unreadable or is rm'ed
at the wrong moment will be ignored instead of raising an exception.
It also bums out a couple of unnecessary lines of code (the explicit
".close()" call).

darcs-hash:20070822172957-92b7f-c1081151636e689d99a8ec340e71f2d6fcc7b016.gz