set the zip_safe flag to False
This means that by default the allmydata-tahoe egg will be a directory with a tree of files instead of a zip file containing files.  I prefer the former because it makes it easier for people to hack into it.
Unfortunately the files therein are still going to be .pyc's instead of .py's, if I understand correctly.  I would prefer for them to be .py's.  See also discussion on the distutils-sig mailing list:

http://mail.python.org/pipermail/distutils-sig/2007-July/007827.html

darcs-hash:20070913223755-92b7f-6c2d86c866e61966034de4ce8dd049f7e0ca2e2b.gz