tweaks to build process to support py2exe
py2exe is unable to handle .eggs which are packaged as zip files
in preference it will pull in other versions of libraries if they
can be found in the environment.

this changes causes .eggs to be built as .egg directories, which
py2exe can handle.

darcs-hash:20080110010253-80e44-dc6a2ee8323fa5294c1042ca9a94507298e74ceb.gz