munin stats: suppress output of data more that 5min old
if a node fails to report stats, the natural thing to do in re munin is to
supress the data for that data series.  the previous tahoe-stats would output
whatever data was present in the stats_gatherer's stats.pickle, regardless of
how old.

this change means that if the gatherer hasn't received data within the last
5 min, then no data is reported to munin for that node.

darcs-hash:20080201030423-80e44-de15094c82b9f22a94f4c134ff276dde3490b1df.gz