make the nodeurl start with '127.0.0.1' instead of with 'localhost'
This is because there exist in the wild computers that are misconfigured so that 'localhost' doesn't resolve to 127.0.0.1.  On those computers, using 'localhost' for the nodeurl is a security problem, because the user commonly sends valuable caps to the nodeurl.

darcs-hash:20080311131803-92b7f-73f6b54d6f8bcd2c49df7ae6334abdc162bfb155.gz