webish: fix directory url redirection
when rendering a directory in webish, if the url did not correctly include
a trailing slash '/' then the browser will be redirected to a url which
does.  this causes any relative links (e.g. the 'other representations'
which are links to files relative to the current directory) to be correctly
intrpreted as relative to the directory in question by the browser.

uses twisted's addSlash mechanism, which is designed for this purpose :-)

darcs-hash:20080611194628-80e44-8bdac4810d86ced98b7258d9a81ae1cec4489b15.gz